Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -5,8 +5,8 @@ char* welcome() {
 }
 
 int main() {
-    int a,b,c;
-    scanf("%d %d %d",&a,&b,&c);
+    float a,b,c;
+    scanf("%f %f %f",&a,&b,&c);
     float Average;
     Average=(a+b+c)/3;
     printf("Average: %.2f",Average);
